requirements.txt
streamlit
pdfplumber
python-dotenv
gTTS
requests
scipy
